test : extract tests from src folder
