+++ Backend/.vscode/settings.json
@@ -3,4 +3,9 @@
     "flake8.args": [
         "--max-line-length=88"
     ],
+    "python.testing.pytestArgs": [
+        "tests"
+    ],
+    "python.testing.unittestEnabled": false,
+    "python.testing.pytestEnabled": true
 }
