+++ Backend/src/model/Playlist.py
@@ -2,7 +2,7 @@ import json
 from dataclasses import dataclass
 from datetime import datetime
 
-import services.song_service as song_service
+import src.services.song_service as song_service
 
 
 @dataclass
