+++ Backend/src/routers/login.py
@@ -2,7 +2,7 @@ import json
 from datetime import datetime, timedelta, timezone
 from typing import Annotated
 
-import services.security_service as security_service
+import src.services.security_service as security_service
 from fastapi import APIRouter, Depends
 from fastapi.responses import Response
 from fastapi.security import OAuth2PasswordRequestForm
