+++ Backend/tests/test_API/api_test_artist.py
@@ -1,5 +1,5 @@
 from fastapi.testclient import TestClient
-from main import app as app
+from src.main import app as app
 
 client = TestClient(app)
