+++ Backend/tests/test_all_users.py
@@ -1,7 +1,7 @@
 import json
 
 import pytest
-from model.UserType import User_Type
+from src.model.UserType import User_Type
 from test_API.api_all_users import (
     delete_playlist_saved,
     patch_history_playback,
