+++ Backend/tests/test_artist.py
