+++ Backend/tests/test_song.py
@@ -1,7 +1,7 @@
 import json
 
 import pytest
-from model.Genre import Genre
+from src.model.Genre import Genre
 from test_API.api_test_artist import create_artist, delete_artist, get_artist
 from test_API.api_test_song import (
     create_song,
